Update 'historical record' use case note
Shade down the claims on 'writing to the object store first.' Recent
revisions removed that structure. Instead, we write to Kafka first,
which has some downsides (might publish a message and then the
archivist fails - boo!) but also some upsides (will always be sure
that the archived data is stuff we have actually published - yay!).